Extract suitemaking out of driver
This creates a new module, suitemaker.py, which contains the
code used to create a suite.GabbiSuite, which contains
case.HTTPTestCase tests.

This limits driver.py to being mostly code that is necessary to
assemble the necessary information to create tests that are runnable
by python test harnesses. This is distinguished from runner.py
which assembles the necessary information to both create and _run_
tests itself.

The idea here is that if the interface presented by test_suite_from_dict
(both signature and that it returns a GabbiSuite) is preserved, then
anything in suitemaker.py is welcome to change. The signature on
test_suite_from_dict is the definition of what a caller (driver or runner
or whatever else) needs to have learned:

loader: the unittest.TestLoader which we might be able to default
        out of existence
test_base_name: names help with test grouping so useful
suite_dict: the main piece of the pie, the dict
test_directory: the place where `data` can load files from
host: the host any unqualified test goes to
port: the port any unqualified test goes to
fixture_module: the sole module from which fixtures come
intercept: the wsgi app to intercept if any
prefix: the url prefix

host and intercept are mutually exclusive

One can image it might be possible to replace host, port, prefix
with a URL instead. If that were done than host and url would be mutually
exclusive.

If the signature of test_suite_from_dict changes then stuff in both driver
and runner, in how they assemble the required info, would need to change.

Note: For reasons unclear these changes have exposed some state management
issues in test_runner, noted in a NOTE.
