+++ gabbi/runner.py
@@ -21,8 +21,9 @@ from six.moves.urllib import parse as urlparse
 import yaml
 
 from gabbi import case
-from gabbi import driver
+from gabbi import handlers
 from gabbi.reporter import ConciseTestRunner
+from gabbi import suitemaker
 
 
 def run():
@@ -117,7 +118,7 @@ def run():
     for import_path in args.response_handlers or []:
         for handler in load_response_handlers(import_path):
             custom_response_handlers.append(handler)
-    for handler in driver.RESPONSE_HANDLERS + custom_response_handlers:
+    for handler in handlers.RESPONSE_HANDLERS + custom_response_handlers:
         handler(case.HTTPTestCase)
 
     data = yaml.safe_load(sys.stdin.read())
@@ -129,10 +130,10 @@ def run():
         else:
             data['defaults'] = {'ssl': True}
     loader = unittest.defaultTestLoader
-    suite = driver.test_suite_from_dict(loader, 'input', data, '.',
-                                        host, port, None, None,
-                                        prefix=prefix)
-    result = ConciseTestRunner(verbosity=2, failfast=args.failfast).run(suite)
+    test_suite = suitemaker.test_suite_from_dict(
+        loader, 'input', data, '.', host, port, None, None, prefix=prefix)
+    result = ConciseTestRunner(
+        verbosity=2, failfast=args.failfast).run(test_suite)
     sys.exit(not result.wasSuccessful())
