+++ gabbi/tests/test_driver.py
@@ -70,99 +70,3 @@ class DriverTest(unittest.TestCase):
         first_test = suite._tests[0]._tests[0]
         full_url = first_test._parse_url(first_test.test_data['url'])
         self.assertEqual('http://localhost:8001/', full_url)
-
-    def test_tests_key_required(self):
-        test_yaml = {'name': 'house', 'url': '/'}
-
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertEqual('malformed test file, "tests" key required',
-                         str(failure.exception))
-
-    def test_upper_dict_required(self):
-        test_yaml = [{'name': 'house', 'url': '/'}]
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertEqual('malformed test file, invalid format',
-                         str(failure.exception))
-
-    def test_inner_list_required(self):
-        test_yaml = {'tests': {'name': 'house', 'url': '/'}}
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertIn('test chunk is not a dict at',
-                      str(failure.exception))
-
-    def test_name_key_required(self):
-        test_yaml = {'tests': [{'url': '/'}]}
-
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertEqual('Test name missing in a test in foo.',
-                         str(failure.exception))
-
-    def test_url_key_required(self):
-        test_yaml = {'tests': [{'name': 'missing url'}]}
-
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertEqual('Test url missing in test foo_missing_url.',
-                         str(failure.exception))
-
-    def test_unsupported_key_errors(self):
-        test_yaml = {'tests': [{
-            'url': '/',
-            'name': 'simple',
-            'bad_key': 'wow',
-        }]}
-
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertIn("Invalid test keys used in test foo_simple:",
-                      str(failure.exception))
-
-    def test_method_url_pair_format_error(self):
-        test_yaml = {'defaults': {'GET': '/foo'}, 'tests': []}
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertIn('"METHOD: url" pairs not allowed in defaults',
-                      str(failure.exception))
-
-    def test_method_url_pair_duplication_format_error(self):
-        test_yaml = {'tests': [{
-            'GET': '/',
-            'POST': '/',
-            'name': 'duplicate methods',
-        }]}
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertIn(
-            'duplicate method/URL directive in "foo_duplicate_methods"',
-            str(failure.exception)
-        )
-
-    def test_dict_on_invalid_key(self):
-        test_yaml = {'tests': [{
-            'name': '...',
-            'GET': '/',
-            'response_html': {
-                'foo': 'hello',
-                'bar': 'world',
-            }
-        }]}
-
-        with self.assertRaises(driver.GabbiFormatError) as failure:
-            driver.test_suite_from_dict(self.loader, 'foo', test_yaml, '.',
-                                        'localhost', 80, None, None)
-        self.assertIn(
-            "invalid key in test: 'response_html'",
-            str(failure.exception)
-        )
